split interface declaration with implementation transaction
